requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aionotify==0.2.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
cx-Freeze==6.7
idna==3.2
importlib-metadata==4.6.4
janus==0.6.1
loguru==0.5.3
multidict==5.1.0
pyftpdlib==1.5.6
pyserial==3.5
pyserial-asyncio==0.5
PySimpleGUI==4.47.0
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
watchgod==0.7
wikipedia==1.4.0
yarl==1.6.3
zipp==3.5.0
